Parse basic numeric expressions.

+++ src/lib/traits/push.rs
@@ -3,5 +3,5 @@ pub trait Push<T> {
 }
 
 pub trait PushLine<T> {
-  fn push_line(&mut self, t: T, line: i32) -> usize;
+  fn push_line(&mut self, t: T, line: i32) -> u8; 
 }
